+++ com.unity.render-pipelines.lightweight/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [5.3.0] - 2018-11-xx
+### Added
+- LWRP now uses the Unity recorder extension to make it possible to capture the output of Cameras.
 
 ## [5.2.0] - 2018-11-27
 ### Added